fix: AgentConfig name 改为可选字段，默认同 id；id 恢复必填
- AgentConfig.name: String → Option<String> + #[serde(default)]
- ResolvedAgentConfig::from_single/merge: name None/空串兜底为 id
- ResolvedAgentConfig: id 终态校验（空 id 返回 ConfigError::MissingId）
- AgentDirectoryProvider: config.id 为空时用目录名注入，不一致时 WARN
- From<AgentConfig> → TryFrom<AgentConfig>
- 新增 7 个 UT 覆盖全部新逻辑
- 拆分 directory_tests.rs（文件行数合规）

Source: issue #834
Type: fix
